requirements.txt
streamlit==1.32.0
azure-ai-documentintelligence==1.0.0
azure-core>=1.30.0
python-dotenv==1.0.0
crewai==0.11.2
langchain==0.1.0
langchain-openai==0.0.5
pydantic==2.5.0
pydantic-core==2.14.1
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
pdf2image==1.16.3
